Unit conversion challenge
A small challenge to create an app that supports multiple conversion types (length, temperature, and time in my version) with a variety of input and out units for each. The user selects a conversion type then enters a value. As the input and output units are selected, the converted value is recalculated and displayed.